mm: huge_memory: a new debugfs interface for splitting THP tests
We did not have a direct user interface for splitting the compound page
backing a THP and there is no need unless we want to expose the THP
implementation details to users.  Make <debugfs>/split_huge_pages accept a
new command to do that.

By writing "<pid>,<vaddr_start>,<vaddr_end>" to
<debugfs>/split_huge_pages, THPs within the given virtual address range
from the process with the given pid are split.  It is used to test
split_huge_page function.  In addition, a selftest program is added to
tools/testing/selftests/vm to utilize the interface by splitting PMD THPs
and PTE-mapped THPs.

This does not change the old behavior, i.e., writing 1 to the interface to
split all THPs in the system.

Link: https://lkml.kernel.org/r/20210319005219.13595-1-zi.yan@sent.com
Signed-off-by: Zi Yan <ziy@nvidia.com>
Cc: "Kirill A. Shutemov" <kirill.shutemov@linux.intel.com>
Cc: Shuah Khan <shuah@kernel.org>
Cc: John Hubbard <jhubbard@nvidia.com>
Cc: Sandipan Das <sandipan@linux.ibm.com>
Cc: David Hildenbrand <david@redhat.com>
Cc: Yang Shi <shy828301@gmail.com>
Cc: Mika Penttila <mika.penttila@nextfour.com>
Cc: Zi Yan <ziy@nvidia.com>
Cc: Mike Kravetz <mike.kravetz@oracle.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>
Signed-off-by: Stephen Rothwell <sfr@canb.auug.org.au>