lib/hexdump: add a raw pointer printing format for slub debugging
Patch series "slub: Print non-hashed pointers in slub debugging".

I was doing some debugging recently and noticed that my pointers were
being hashed while slub_debug was on the kernel commandline.  Let's make
the prints in here meaningful in that case by pushing %px throughout.

Alternatively, we could force on no_hash_pointers if slub_debug is on the
commandline.  Maybe that would be better?

The final patch is just something else I noticed while looking at the
code.  The message argument is never used so the debugging messages are
not as clear as they could be.

This patch (of 3):

We want to get actual pointer addresses when we're looking at slub
debugging reports.  Add another prefix format specifier that says we want
raw pointer addresses, i.e.  %px, in the printk format.

Link: https://lkml.kernel.org/r/20210520013539.3733631-1-swboyd@chromium.org
Link: https://lkml.kernel.org/r/20210520013539.3733631-2-swboyd@chromium.org
Signed-off-by: Stephen Boyd <swboyd@chromium.org>
Cc: Christoph Lameter <cl@linux.com>
Cc: Pekka Enberg <penberg@kernel.org>
Cc: David Rientjes <rientjes@google.com>
Cc: Joonsoo Kim <iamjoonsoo.kim@lge.com>
Cc: Vlastimil Babka <vbabka@suse.cz>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>
Signed-off-by: Stephen Rothwell <sfr@canb.auug.org.au>