drm/amdgpu/smu7: split PM logging start and sampling
We need to use the PM log to get the average power.  Start
the logging at asic init time and then just query the
samples at runtime to get the average power rather than
restarting the sampling and waiting for the values to
accumulate on every sensor query.  This should reduce
mutex contention in the driver when applications poll
the hwmon data in tight loops.

Bug: https://gitlab.freedesktop.org/drm/amd/-/issues/1660
Signed-off-by: Alex Deucher <alexander.deucher@amd.com>